fix(namings): renamed and moved some .py files around for consistent naming
